Refactor no fluxo de exibição do detalhamento do jogo

+++ src/main/java/br/com/fatec/chopperhousegames/core/domain/service/JogoService.java
@@ -12,5 +12,5 @@ public interface JogoService {
     Jogo ativar(Jogo jogo);
     List<Jogo> listar();
     List<Jogo> listarAtivos();
-    Jogo buscarById(Integer id);
+    Jogo buscarById(Long id);
 }
