+++ src/main/java/br/com/fatec/chopperhousegames/inbound/facade/impl/JogoFacadeImpl.java
@@ -33,7 +33,7 @@ public class JogoFacadeImpl implements JogoFacade {
     }
 
     @Override
-    public JogoDTO buscarById(Integer id) {
+    public JogoDTO buscarById(Long id) {
         return mapper.toJogoDTO(jogoService.buscarById(id));
     }
